server/requirements.txt
charset-normalizer==3.1.0
dicttoxml==1.7.16
google-api-core==2.11.0
jsonpickle==3.0.1
lxml==4.9.2
msgpack==1.0.5
protobuf==4.22.3
PyYAML==6.0
requests==2.28.2
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.15
xml-marshaller==1.0.2
xmltodict==0.13.0
dataclasses-avroschema==0.42.0

